refactor: sparate post edit page code
